refactor: rename components/elements file
