+++ src/components/Saider/Elements/Lable.ts
@@ -1,9 +1,10 @@
 import styled from 'styled-components';
 
-const Label = styled.label<{ staus: boolean; message: string }>`
+const Label = styled.label<{ status: boolean; message: string }>`
   position: relative;
+  color: ${(props) => (props.status ? 'inherit' : 'red')};
   &:after {
-    content: '*${(props) => (props.staus ? '' : props.message)}';
+    content: '*${(props) => (props.status ? '' : props.message)}';
     position: absolute;
     left: 0;
     bottom: -12px;
